refactor: mover documentos a modulo de categorias
